+++ src/Inputs/SearchUtils.js
@@ -118,6 +118,16 @@ export const CONDITIONS = {
     minFields: 0,
     invalidInputTypes: [...ALL_BUT_DATES]
   },
+  'this year': {
+    maxFields: 0,
+    minFields: 0,
+    invalidInputTypes: [...ALL_BUT_DATES]
+  },
+  'last year': {
+    maxFields: 0,
+    minFields: 0,
+    invalidInputTypes: [...ALL_BUT_DATES]
+  },
   'year to date': {
     maxFields: 0,
     minFields: 0,
@@ -129,6 +139,11 @@ export const CONDITIONS = {
     invalidInputTypes: [...ALL_BUT_DATES]
   },
   'fiscal year': {
+    maxFields: 0,
+    minFields: 0,
+    invalidInputTypes: [...ALL_BUT_DATES]
+  },
+  'last fiscal year': {
     maxFields: 0,
     minFields: 0,
     invalidInputTypes: [...ALL_BUT_DATES]